Moved aware_datetime in conftest
