+++ procrastinate/utils.py
@@ -242,26 +242,6 @@ def utcnow() -> datetime.datetime:
     return datetime.datetime.now(tz=datetime.timezone.utc)
 
 
-def aware_datetime(
-    year: int,
-    month: int,
-    day: int,
-    hour: int = 0,
-    minute: int = 0,
-    second: int = 0,
-    microsecond: int = 0,
-    tz_offset: Optional[int] = None,
-) -> datetime.datetime:
-    tzinfo = (
-        datetime.timezone(datetime.timedelta(hours=tz_offset))
-        if tz_offset
-        else datetime.timezone.utc
-    )
-    return datetime.datetime(
-        year, month, day, hour, minute, second, microsecond, tzinfo=tzinfo
-    )
-
-
 def parse_datetime(raw: str) -> datetime.datetime:
     try:
         # this parser is the stricter one, so we try it first
